Safari-609.1.11.1/LayoutTests/css3/flexbox/box-sizing-min-max-sizes.html

<!DOCTYPE html>
<html>
<style>
.flexbox {
    background-color: teal;
}
.flexbox > div {
    box-sizing: border-box;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="flexbox">
    <div style="border: 5px solid orange; padding: 5px; max-width: 30px; width: 40px;" data-expected-width=30></div>
    <div style="border: 5px solid orange; padding: 5px; min-width: 30px; width: 10px;" data-expected-width=30></div>
</div>

<div class="flexbox column">
    <div style="border: 5px solid orange; padding: 5px; max-height: 30px; height: 40px;" data-expected-height=30></div>
    <div style="border: 5px solid orange; padding: 5px; min-height: 30px; height: 10px;" data-expected-height=30></div>
</div>

</body>
</html>
